requirements.txt
appnope==0.1.0
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coverage==4.5.2
cycler==0.10.0
decorator==4.3.0
dj-static==0.0.6
Django==2.1.4
django-bmemcached==0.2.3
django-crispy-forms==1.7.2
django-request-logging==0.6.6
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
Faker==1.0.2
flake8==3.6.0
glfw==1.7.0
guess-indian-gender==1.0.1
idna==2.8
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
kiwisolver==1.0.1
matplotlib==3.0.2
mccabe==0.6.1
mock==2.0.0
more-itertools==5.0.0
mysqlclient==1.4.1
nltk==3.4
numpy==1.15.4
oauthlib==2.1.0
pandas==0.23.4
parso==0.3.1
pbr==5.1.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.8.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.3.1
PyJWT==1.7.1
PyOpenGL==3.1.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==4.0.2
pytest-cov==2.6.0
pytest-django==3.4.4
python-binary-memcached==0.28.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python-memcached==1.59
python-memcached-stats==0.1
pytz==2018.7
pyxel==1.0.0
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
sentry-sdk==0.6.4
singledispatch==3.4.0.3
six==1.12.0
sounddevice==0.3.12
soupsieve==1.8
static3==0.7.0
text-unidecode==1.2
textblob==0.15.2
traitlets==4.3.2
tweepy==3.7.0
typing==3.6.6
uhashring==1.1
urllib3==1.24.1
wcwidth==0.1.7
wordcloud==1.5.0